Move inference core to ty crate
