+++ crates/engine/body-ir/src/resolution/infer/mod.rs
@@ -1,13 +1,9 @@
 //! Body-local inference facts used before writing resolved `Ty` values.
 //!
-//! The persisted Body IR model stores ordinary `Ty` facts. This module keeps inference variables
-//! in a transient table so later resolver phases can preserve relationships such as `Vec<?T>`
-//! until local evidence solves `?T`.
+//! The persisted Body IR model stores ordinary `Ty` facts. This module only maps body expression
+//! and binding slots to the transient inference table owned by `rg_ty`.
 
 mod context;
-mod instantiate;
-mod model;
-mod table;
 
 pub(super) use context::BodyInferenceCtx;
