+++ crates/engine/ir-model/src/ids/mod.rs
@@ -2,3 +2,5 @@ pub(crate) mod body;
 pub(crate) mod def_map;
 pub mod identity;
 pub(crate) mod semantic;
+
+pub use rg_parse::TargetId;
