+++ crates/engine/ir-model/src/lib.rs
@@ -19,6 +19,7 @@ pub use self::hir::body::{
     RecordExprSpread, RecordFieldSyntax, RecordPatField, ScopeData, StmtData, StmtKind,
 };
 pub use self::ids::{
+    TargetId,
     body::{BindingId, BodyBindingRef, BodyId, BodyRef, ExprId, PatId, ScopeId, StmtId},
     def_map::{
         DefId, DefMapRef, ImportId, ImportRef, LocalDefId, LocalDefRef, LocalImplId, LocalImplRef,
